Added syllable counter (WiP)
Looking to add support for more readability indexes, because I don't trust the ARI, and more data is good data. Problem is, most require syllable counts in their calculation, which requires a means of counting them.
This is not easy to do algorithmically, and is impossible to do with 100% accuracy, because English is a major cluster**** of a language. As with most things in this language, syllable counts follow a simple set of rules which holds true at best 30% of the time. As such, I'm hard-coding in a lot of the more common exceptions.
I'm not happy with this method; it's very kludgy. It is giving me good results, however, in some ways better results than many existing syllable counters. I'm hoping to be able to condense it into something better once I've completed this iteration.
-RB